BREAKING: Consolidate beam transforms into I/O-scoped modules (0.14.0)
Per-direction modules (read_from_bigquery, write_to_json, etc.) are replaced
by I/O-scoped modules that group related transforms together:

  read_from_bigquery.py + bigquery.py → bigquery.py
  read_from_json.py + write_to_json.py → json.py
  read_matching_avro_files.py → avro.py
  apply_sliding_windows.py → windows.py
  group_by.py + sample_and_log.py → core.py

All public names remain importable from gfw.common.beam.transforms.
Direct imports from the old module paths (e.g.
gfw.common.beam.transforms.read_from_bigquery) will break and must be
updated to the new paths.
