+++ tests/beam/transforms/test_read_from_json.py
@@ -5,7 +5,7 @@ import pytest
 from apache_beam.testing.test_pipeline import TestPipeline
 from apache_beam.testing.util import assert_that, equal_to
 
-from gfw.common.beam.transforms.read_from_json import ReadFromJson
+from gfw.common.beam.transforms.json import ReadFromJson
 from gfw.common.io import json_save
